0.5 : - Map component and Fight component are now connected; if the player interacts with a life fountain, the player is healed, and if it interacts with a monster, the Fight component is triggered and the player can fight the monster to defeat it - Removed unnecessary classes, properties and methods
