--- FightManager.cs
@@ -1,1 +1,0 @@
-using JDR;
