+++ Models/LogService.cs
